<⚡️ perf>[client: patrol-view]:{use context for patrol Data}
