+++ client/app/[locale]/patrol/[id]/layout.tsx
@@ -1,0 +1,18 @@
+'use client'
+
+import { PatrolProvider, usePatrol } from '@/context/patrol-context';
+
+export default function PatrolLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+
+    return (
+        <PatrolProvider>
+            <div className="flex flex-col gap-4 px-6 py-4">
+                {children}
+            </div>
+        </PatrolProvider>
+    );
+}
