+++ client/app/[locale]/patrol/layout.tsx
@@ -1,14 +1,16 @@
+'use client'
 import Header from "@/components/header";
 
 export default function PatrolLayout({
-    children,
-  }: Readonly<{
-    children: React.ReactNode;
-  }>) {
-    return (
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
     <>
-      <Header variant="inspector"/>
-      {children}
+        <Header variant="inspector" />
+        {children}
+
     </>
-    );
-  }
+  );
+}
